Refactor code so maxPages is obtained on fetch
The maximum amount calc of pages used by <Pagination /> was hardcoded
inside the component. This is a problem since if the API returns a
different amount of products than 194 (hardcoded value) it may break.

Now fetching the products also returns the total amount of products
in the API, and the calc is done in the custom hook `useProducts`.
Also exported the amount of products being fetch from the services
to keep all in sync (considering moving that constant elsewhere to
avoid having UI logic in data fetching functions)